refactor: rename `x_idefics_2_y` files to `x_idefics2_y`
