+++ colpali_engine/collators/visual_retriever_collator.py
@@ -2,7 +2,7 @@ from typing import Any, Dict, List, Union, cast
 
 from PIL.Image import Image
 
-from colpali_engine.models.idefics_2 import ColIdefics2Processor
+from colpali_engine.models.idefics2 import ColIdefics2Processor
 from colpali_engine.models.paligemma import ColPaliProcessor
 from colpali_engine.utils.processing_utils import BaseVisualRetrieverProcessor
