+++ colpali_engine/models/__init__.py
@@ -1,3 +1,3 @@
-from .idefics_2 import BiIdefics2, ColIdefics2, ColIdefics2Processor
+from .idefics2 import BiIdefics2, BiIdefics2Processor, ColIdefics2, ColIdefics2Processor
 from .paligemma import BiPali, BiPaliProcessor, BiPaliProj, ColPali, ColPaliProcessor
 from .qwen2 import BiQwen2, BiQwen2Processor, ColQwen2, ColQwen2Processor
