+++ colpali_engine/models/idefics2/__init__.py
@@ -1,0 +1,2 @@
+from .biidefics2 import BiIdefics2, BiIdefics2Processor
+from .colidefics2 import ColIdefics2, ColIdefics2Processor
