+++ colpali_engine/models/idefics2/biidefics2/modeling_biidefics2.py
