+++ colpali_engine/models/idefics2/colidefics2/modeling_colidefics2.py
