+++ colpali_engine/models/idefics2/colidefics2/processing_colidefics2.py
