--- colpali_engine/models/idefics_2/__init__.py
@@ -1,2 +1,0 @@
-from .biidefics_2 import BiIdefics2
-from .colidefics_2 import ColIdefics2, ColIdefics2Processor
