--- colpali_engine/models/idefics_2/biidefics_2/__init__.py
@@ -1,1 +1,0 @@
-from .modeling_biidefics_2 import BiIdefics2
