--- colpali_engine/models/idefics_2/colidefics_2/__init__.py
@@ -1,2 +1,0 @@
-from .modeling_colidefics_2 import ColIdefics2
-from .processing_colidefics_2 import ColIdefics2Processor
